+++ beets/library.py
@@ -157,6 +157,7 @@ PF_KEY_DEFAULT = 'default'
 log = logging.getLogger('beets')
 if not log.handlers:
     log.addHandler(logging.StreamHandler())
+log.propagate = False  # Don't propagate to root handler.
 
 # A little SQL utility.
 def _orelse(exp1, exp2):
@@ -1001,7 +1002,7 @@ class Library(BaseLibrary):
             self.path = bytestring_path(normpath(path))
         self.directory = bytestring_path(normpath(directory))
         self.path_formats = path_formats
-        self.art_filename = bytestring_path(art_filename)
+        self.art_filename = art_filename
         self.replacements = replacements
 
         self._memotable = {}  # Used for template substitution performance.
@@ -1159,6 +1160,9 @@ class Library(BaseLibrary):
             # Outputting Unicode.
             extension = extension.decode('utf8', 'ignore')
         subpath += extension.lower()
+
+        # Truncate too-long components.
+        subpath = util.truncate_path(subpath, pathmod)
 
         if fragment:
             return subpath
@@ -1559,9 +1563,19 @@ class Album(BaseAlbum):
         """
         image = bytestring_path(image)
         item_dir = item_dir or self.item_dir()
+
+        if not isinstance(self._library.art_filename,Template):
+            self._library.art_filename = Template(self._library.art_filename)
+
+        subpath = util.sanitize_path(util.sanitize_for_path(
+            self.evaluate_template(self._library.art_filename)
+        ))
+        subpath = bytestring_path(subpath)
+
         _, ext = os.path.splitext(image)
-        dest = os.path.join(item_dir, self._library.art_filename + ext)
-        return dest
+        dest = os.path.join(item_dir, subpath + ext)
+
+        return bytestring_path(dest)
 
     def set_art(self, path, copy=True):
         """Sets the album's cover art to the image at the given path.